Replaced sets with ordered sets.
The basic paradigm of simply using sets is great, the mecanism of
sorting the set before adding it tothe playlist has limitations,
especially when generating random playlists. I replaced sets with
ordered sets so that the set can be shuffled and the original order of
the album can be kept without sorting the set.

- Added the OrderedSet class
- The MPDHelper.sort class not only sorted the given list of songs, but also filtered out the songs that are not in the collections. Renamed it to filter and kept only the second behaviour.